requirements/flake8.txt
bandit==1.6.2
black==19.10b0
django-stubs==1.5.0
djangorestframework-stubs==1.2.0
flake8==3.8.3
flake8-blind-except==0.1.1
flake8-awesome==1.3.0
flake8-bugbear==20.1.4
flake8-commas==2.0.0
flake8-debugger==3.2.1
flake8-docstrings==1.5.0
flake8-pep3101==1.3.0
flake8-quotes==3.2.0
flake8-rst-docstrings==0.0.13
flake8-string-format==0.3.0
lizard==1.17.4
mccabe==0.6.1
mypy==0.782
pep257==0.7.0
pep8==1.7.1
pep8-naming==0.11.1
pip-check==2.6
pygments==2.6.1
radon==4.1.0
yamllint==1.24.2
